requirements.txt
Flask==2.0.1
RPi.GPIO==0.7.0
pyngrok==5.1.0
Flask-BasicAuth==0.2.0